test(unit): converted the project-type scaffolder tests to vitest
for #452
